+++ handshake/crypto_setup_server_test.go
@@ -184,8 +184,11 @@ var _ = Describe("Server Crypto Setup", func() {
 		Expect(err).NotTo(HaveOccurred())
 		version = protocol.SupportedVersions[len(protocol.SupportedVersions)-1]
 		supportedVersions = []protocol.VersionNumber{version, 98, 99}
-		cpm = NewConnectionParamatersManager(protocol.PerspectiveServer, protocol.VersionWhatever,
+		cpm = NewConnectionParamatersManager(
+			protocol.PerspectiveServer,
+			protocol.VersionWhatever,
 			protocol.DefaultMaxReceiveStreamFlowControlWindowServer, protocol.DefaultMaxReceiveConnectionFlowControlWindowServer,
+			protocol.DefaultIdleTimeout,
 		)
 		csInt, err := NewCryptoSetup(
 			protocol.ConnectionID(42),